pitchwise-backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
SQLAlchemy==2.0.36
pymysql==1.1.1
cryptography==44.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Validation
pydantic==2.10.5
pydantic[email]==2.10.5

# Environment Variables
python-dotenv==1.0.1

# CORS
fastapi-cors==0.0.6
